pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.3.21",
    "db-sqlite3>=0.0.1",
    "fastapi>=0.124.4",
    "google-adk>=1.18.0",
    "mcp>=1.24.0",
    "openai>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
